requirements.txt
Django==4.2.9
djangorestframework==3.14.0
python-decouple==3.8